+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/mock/msw/handlers.ts
@@ -1,0 +1,25 @@
+import { http, HttpResponse } from 'msw';
+import { mockData } from '../data';
+import { getMellomlagringHandlers } from './mellomlagringHandlers';
+
+export const getHandlers = () => [
+    http.get('**/oppslag/soker', () => {
+        return HttpResponse.json(mockData.søker);
+    }),
+    http.get('**/oppslag/barn', () => {
+        return HttpResponse.json(mockData.barn);
+    }),
+    http.post('**/omsorgspenger-midlertidig-alene/innsending', () => {
+        return HttpResponse.json({});
+    }),
+    http.post('**/vedlegg', () => {
+        return HttpResponse.json({}, { headers: { location: '/vedlegg/123' } });
+    }),
+    http.delete('**/vedlegg/*', () => {
+        return HttpResponse.json({});
+    }),
+    ...getMellomlagringHandlers('OMSORGSPENGER_MIDLERTIDIG_ALENE', 'mellomlagring/OMSORGSPENGER_MIDLERTIDIG_ALENE'),
+
+    // Stopp alle andre kall (Dekoratøren og andre tredjepartstjenester)
+    http.all('*', () => new HttpResponse(null, { status: 200 })),
+];