+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/tailwind.config.ts
@@ -1,0 +1,12 @@
+import type { Config } from 'tailwindcss';
+
+const config: Config = {
+    content: ['src/**/*.{tsx,ts}', '../../packages/sif-common-core-ds/**/*.{tsx, ts}'],
+    theme: {
+        extend: {},
+    },
+    plugins: [],
+    presets: [require('@navikt/ds-tailwind')],
+};
+
+export default config;