+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/vite.config.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /// <reference types="vitest" />
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
@@ -7,7 +6,7 @@ import checker from 'vite-plugin-checker';
 export default defineConfig({
     plugins: [
         react({
-            include: '**/*.{jsx,tsx}',
+            include: '**/*.{tsx}',
         }),
         checker({ typescript: true }),
         {
@@ -20,11 +19,23 @@ export default defineConfig({
     resolve: {},
     build: {
         sourcemap: true,
+        target: 'esnext',
+        rollupOptions: {
+            output: {
+                manualChunks: {
+                    vendor: ['react', 'react-dom', 'react-intl'],
+                    router: ['react-router-dom'],
+                    forms: ['formik'],
+                    utils: ['lodash', 'date-fns', 'dayjs', 'axios', 'uuid', 'zod'],
+                    navikt: ['@navikt/ds-react', '@navikt/ds-icons', '@navikt/ds-css', '@navikt/ds-tailwind'],
+                },
+            },
+        },
     },
     css: {
         preprocessorOptions: {
             scss: {
-                api: 'modern-compiler',
+                api: 'modern-compiler', // or "modern", "legacy"
             },
         },
     },