+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/storybook/stories/steps/Fravær.stories.tsx
@@ -1,5 +1,4 @@
 import { Meta, StoryFn } from '@storybook/react-vite';
-import * as React from 'react';
 import FraværStep from '../../../app/søknad/steps/fravær/FraværStep';
 import { withAmplitudeProvider } from '../../decorators/withAmplitudeProvider';
 import { withFormikWrapper } from '../../decorators/withFormikWrapper';