+++ apps/pleiepenger-sykt-barn/src/storybook/data/søkerMock copy 2.ts
@@ -1,9 +1,0 @@
-import { Søker } from '../../../src/app/types/Søker';
-
-export const søkerMock: Søker = {
-    fornavn: 'Ola',
-    etternavn: 'Nordmann',
-    fødselsnummer: '12345678901',
-    mellomnavn: '',
-    kjønn: 'mann',
-};