+++ packages/k9-sak-innsyn-k9-sak-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.0.9
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + sentry 10
+
+## 0.0.8
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.0.7
 
 ### Patch Changes