+++ packages/ung-deltakelse-opplyser-api-deltaker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/ung-deltakelse-opplyser-api
 
+## 0.0.18
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + sentry 10
+
+## 0.0.17
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.0.16
 
 ### Patch Changes